Update setup.py to allow packages to update within minor releases
This commit updates cti-toolkit to support STIX 1.2+, but wont update to STIX 2.0
It also allows libtaxii to update, but wont update to libtaxii 2.0. And it does
the same for CybOX 2.1 (but not to Cybox 3.0). I've also taking the liberty to
do the same for stix-ramrod, lxml, mixbox and pymisp.

These changes will allow maintenance releases and minor updates to automatically
be applied without affecting cti-toolkit. This is especially important for
STIX content generation, as many TIPs have been updated with STIX 1.,2 and now
can only generate STIX 1.2 content. Without this change CTI-Toolkit cannot
process feeds from these STIX 1.2 exporting TIPs.