Improved test configuration
Modified tox and pytest configuration to make it easier to add python3
and flake8 (PEP8) tests in the future. Also moved requirements into
requirements.tx and requirements-dev.txt to allow for reuse in both
packaging and running tests. Added MANIFEST.in to specify exactly
what goes in the packages generated. Extended .gitignore file to
ignore the new coverage output, and more common directories and file
types (such as use of the IntelliJ IDEA IDE.<F3>