Added autodetection of STIX content and new Report object
STIX 1.2 moved the STIX Package title and description to a new
Report object. This commit updates CTI-toolkit to support this new
location in addition to the original location. This is all handled
transparently to the user, as part of the STIX import process.

TOX now tests STIX 1.1.1 and STIX 1.2 content, and passes all tests