Modified the MISP tests to reflect MISP 2.4.81
Updated the MISP tests so that they worked with the later
versions of MISP. Emulated the attribute adding with a
httpretty callback to accurately reflect how the MISP
server responds. Also Made the MISP emulated version respond
with the version of PyMISP so that they are always in lockstep
and the PyMISP won't error (as we're testing that PyMISP is
sending commands, we're not actually emulated an entire
MISP server :).