Changed Snort DNS output so alerts when DNS lookup rather than HTTP
The main change of this commit is to modifiy the behaviour so that
the snort alerts for domains are associated with the DNS lookup
behaviour. This will catch the traffic earlier than just looking at
HTTP, and will also allow us to alert if the site is using HTTPS
(because it still needs to do DNS lookup locally!)

A secondary change is to update the installation docs so that they
contain information on how to do a local development install if
anyone else wants to extend or enhance the cti-toolkit.

And lastly, I've added my name to the main README.rst so that
my contributions are documented.